refactor: move PerpsPillItem and ExplorePill components to UI folder
