+++ app/components/UI/Perps/components/PerpsPillItem/index.ts
@@ -1,0 +1,1 @@
+export { default as PerpsPillItem } from './PerpsPillItem';
