+++ app/components/UI/Trending/components/ExplorePill/ExplorePill.test.tsx
