+++ app/components/UI/Trending/components/ExplorePill/index.ts
@@ -1,0 +1,2 @@
+export { default as ExplorePill } from './ExplorePill';
+export type { ExplorePillProps } from './ExplorePill';
