+++ app/components/UI/Trending/components/PillScrollList/index.ts
@@ -1,0 +1,2 @@
+export { default as PillScrollList } from './PillScrollList';
+export type { PillScrollListProps } from './PillScrollList';
