+++ app/components/UI/Trending/components/SectionPillsSkeleton/index.ts
@@ -1,0 +1,3 @@
+export { default as SectionPillsSkeleton } from './SectionPillsSkeleton';
+/** @deprecated Use `SectionPillsSkeleton`. Kept for TrendingView feed compatibility. */
+export { default as CryptoMoversSkeleton } from './SectionPillsSkeleton';
